Fix a couple of clang-tidy warnings
