+++ examples/json/sampleRandLevel.json
@@ -15,9 +15,9 @@
     "make_graphml_open_abc_d": true,
     "create_id_directories": true,
 
-    "min_level": 1000,
-    "max_level": 1000,
-    "min_elem": 1000,
-    "max_elem": 1000
+    "min_level": 100,
+    "max_level": 100,
+    "min_elem": 100,
+    "max_elem": 100
   }
 ]
